rename daqbox api
